pyproject.toml
[project]
name = "jmodel_desktop"
version = "0.1.0"
description = ""
authors = [
    {name = "King-Navi",email = "168604098+King-Navi@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "=3.10.12"
dependencies = [
    "pyside6 (>=6.10.1,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[project.scripts]
start_app = "jmodel_desktop.src.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.17.0"
ruff = "^0.14.9"

